fix: clarify automation self-archive boundary
